bug: authTokenApi bean冲突 #7831
