+++ src/backend/ci/core/auth/biz-auth/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.tencent.devops.common.auth.MockAuthAutoConfiguration
+com.tencent.devops.auth.common.MockAuthConfiguration
